Update pbardownloader.py
fixed some bugs, add file existing detection

+++ pbardownloader.py
@@ -9,9 +9,6 @@ from contextlib import closing
 
 
 class ProgressBar(object):
-    '''
-    ProgressBar类，简单的下载进度显示: [abcd.exe] downloding 84574.00 KB / 171498.00 KB
-    '''
     def __init__(self, title, count=0.0, run_status=None, fin_status=None, total=100.0, unit='', sep='/',
                  chunk_size=1.0):
         super(ProgressBar, self).__init__()
@@ -40,13 +37,11 @@ class ProgressBar(object):
         if self.count >= self.total:
             end_str = '\n'
             self.status = status or self.fin_status
+        #print self.__get_info() + '\b' #+ end_str
         sys.stdout.write(self.__get_info() + '\r')
 
 
 class DownLoader(object):
-    '''
-    Downloader类，实现下载文件的主要功能，支持断点续传
-    '''
     def __init__(self, url, out=None):
         super(DownLoader, self).__init__()
         self.url = url
@@ -64,7 +59,7 @@ class DownLoader(object):
             self.output = ""
         else:
             pass
-            #netloc = os.path.basename(urlparse.urlparse(self.url).netloc)  
+            #netloc = os.path.basename(urlparse.urlparse(self.url).netloc)  #103.5.126.164
         return self.output
 
     def is_continuable(self):
@@ -88,6 +83,7 @@ class DownLoader(object):
         s = self.detect_directory()
         directory = s + "\\" if s.strip() else s
         #print directory
+
         isTrue, Length = self.is_continuable()
         if os.path.exists(directory + filename):
             current_size = int(os.path.getsize(directory + filename))
@@ -103,20 +99,24 @@ class DownLoader(object):
                 chunk_size = 1024
                 #content_size = int(response.headers['content-length'])
                 content_size = Length
-                print content_size
-                progress = ProgressBar(filename, count=current_size, total=content_size, unit="KB", chunk_size=chunk_size, run_status="downloding", fin_status="downloaded")
-                with open(directory + filename, "ab+") as file:
-                    file.seek(current_size)
-                    file.truncate()
-                    for data in response.iter_content(chunk_size=chunk_size):
-                        file.write(data)
-                        file.flush()
-                        progress.refresh(count=len(data))
-
-                print "Saved in " + directory + filename
+                if current_size >= content_size:
+                    print directory + filename, " already exists!"
+                else:
+                    if current_size != 0:
+                        print ">>> download continue >>>"
+                    progress = ProgressBar(filename, count=current_size, total=content_size, unit="KB", chunk_size=chunk_size, run_status="downloding", fin_status="downloaded")
+                    # chunk_size = chunk_size &lt; content_size and chunk_size or content_size
+                    with open(directory + filename, "ab+") as file:
+                        file.seek(current_size)
+                        file.truncate()
+                        for data in response.iter_content(chunk_size=chunk_size):
+                            file.write(data)
+                            file.flush()
+                            progress.refresh(count=len(data))
+                    print "\nSaved in " + directory + filename
         except KeyboardInterrupt:
             #print "Failed to connect [" + filename + "]"
-            sys.stdout.write("\n" + "^C download paused by user")
+            sys.exit("\n^C captured, download paused by user!")
         except requests.exceptions.ReadTimeout, e:
             print e.message
